Add product info and styling to product listing
The styling to the product listing was barebones. To modernize the
product listing CSS/React bootstrap was used for product cards, buttons
and badges. Furthermore, product info only displayed the slug name and
not any info about the price, availability and description. This has
been fixed by requesting the product info from the backend and
displaying it with the help of CSS/React bootstrap.